requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.0
plotly==5.18.0
scipy==1.12.0
scikit-learn==1.4.0
pyyaml==6.0.1
openpyxl==3.1.2